oauth personal access token fully conformed

+++ routes/api.php
@@ -68,16 +68,15 @@ Route::namespace('Api')->name('api.')->group(function () {
      */
     Route::namespace('\Laravel\Passport\Http\Controllers')->group(function() {
       Route::get('/scopes', 'ScopeController@all')->name('scopes.index')->middleware('auth');
-
     });
 
     /**
      * Personal Access Tokens
      */
-    Route::namespace('\Laravel\Passport\Http\Controllers')->group(function() {
+    Route::namespace('OAuth')->group(function() {
       Route::get('/personal-access-tokens', 'PersonalAccessTokenController@forUser')->name('personal.tokens.index')->middleware('auth');
       Route::post('/personal-access-tokens', 'PersonalAccessTokenController@store')->name('personal.tokens.store')->middleware('auth');
-      Route::delete('/personal-access-tokens/{token_id}')->name('personal.tokens.destroy')->middleware('auth');
+      Route::delete('/personal-access-tokens/{token_id}', 'PersonalAccessTokenController@destroy')->name('personal.tokens.destroy')->middleware('auth');
     });
 
     /**
